+++ cnxarchive/views.py
@@ -85,29 +85,6 @@ def get_latest_version(uuid_):
                 raise httpexceptions.HTTPNotFound()
 
 
-def get_content_metadata(id, version, cursor):
-    """Return metadata related to the content from the database."""
-    # Do the module lookup
-    args = dict(id=id, version=version)
-    # FIXME We are doing two queries here that can hopefully be
-    #       condensed into one.
-    cursor.execute(SQL['get-module-metadata'], args)
-    try:
-        result = cursor.fetchone()[0]
-        # version is what we want to return, but in the sql we're using
-        # current_version because otherwise there's a "column reference is
-        # ambiguous" error
-        result['version'] = result.pop('current_version')
-
-        # FIXME We currently have legacy 'portal_type' names in the database.
-        #       Future upgrades should replace the portal type with a mimetype
-        #       of 'application/vnd.org.cnx.(module|collection|folder|<etc>)'.
-        #       Until then we will do the replacement here.
-        result['mediaType'] = portaltype_to_mimetype(result['mediaType'])
-
-        return result
-    except (TypeError, IndexError,):  # None returned
-        raise httpexceptions.HTTPNotFound()
 
 
 def is_latest(id, version):
@@ -229,13 +206,6 @@ def html_listify(tree, root_ul_element, parent_id=None):
             html_listify(node['contents'], elm, parent_id)
 
 
-def tree_to_html(tree):
-    """Return html list version of book tree."""
-    ul = etree.Element('ul')
-    html_listify([tree], ul)
-    return HTML_WRAPPER.format(etree.tostring(ul))
-
-
 def _get_page_in_book(page_uuid, page_version, book_uuid,
                       book_version, latest=False):
     book_ident_hash = join_ident_hash(book_uuid, book_version)
@@ -268,74 +238,6 @@ def _convert_legacy_id(objid, objver=None):
                 return (None, None)
 
 
-def _get_content_json(ident_hash=None):
-    """Return a content as a dict from its ident-hash (uuid@version)."""
-    request = get_current_request()
-    routing_args = request and request.matchdict or {}
-    settings = get_current_registry().settings
-    if not ident_hash:
-        ident_hash = routing_args['ident_hash']
-    id, version = split_ident_hash(ident_hash)
-
-    as_collated = asbool(request.GET.get('as_collated', True))
-    page_ident_hash = routing_args.get('page_ident_hash', '')
-    if page_ident_hash:
-        try:
-            p_id, p_version = split_ident_hash(page_ident_hash)
-        except IdentHashShortId as e:
-            p_id = get_uuid(e.id)
-            p_version = e.version
-        except IdentHashMissingVersion as e:
-            # page ident hash doesn't need a version
-            p_id = e.id
-            p_version = None
-
-    with psycopg2.connect(settings[config.CONNECTION_STRING]) as db_connection:
-        with db_connection.cursor() as cursor:
-            result = get_content_metadata(id, version, cursor)
-            if result['mediaType'] == COLLECTION_MIMETYPE:
-                # Grab the collection tree.
-                result['tree'] = get_tree(ident_hash, cursor,
-                                          as_collated=as_collated)
-                result['collated'] = as_collated
-                if not result['tree']:
-                    # If collated tree is not available, get the uncollated
-                    # tree.
-                    result['tree'] = get_tree(ident_hash, cursor)
-                    result['collated'] = False
-
-                if page_ident_hash:
-                    for id_ in flatten_tree_to_ident_hashes(result['tree']):
-                        id, version = split_ident_hash(id_)
-                        if id == p_id and (
-                           version == p_version or not p_version):
-                            content = None
-                            if as_collated:
-                                content = get_collated_content(
-                                    id_, ident_hash, cursor)
-                            if content:
-                                result = get_content_metadata(
-                                    id, version, cursor)
-                                result['content'] = content[:]
-                                return result
-                            raise httpexceptions.HTTPFound(request.route_path(
-                                request.matched_route.name,
-                                ident_hash=join_ident_hash(id, version)))
-                    raise httpexceptions.HTTPNotFound()
-            else:
-                # Grab the html content.
-                args = dict(id=id, version=result['version'],
-                            filename='index.cnxml.html')
-                cursor.execute(SQL['get-resource-by-filename'], args)
-                try:
-                    content = cursor.fetchone()[0]
-                except (TypeError, IndexError,):  # None returned
-                    logger.debug("module found, but "
-                                 "'index.cnxml.html' is missing.")
-                    raise httpexceptions.HTTPNotFound()
-                result['content'] = content[:]
-
-    return result
 
 
 def notblocked(page):
@@ -405,49 +307,6 @@ def _get_licenses(cursor):
 #   Views   #
 # ######### #
 
-
-@view_config(route_name='content-json', request_method='GET')
-def get_content_json(request):
-    """Retrieve content as JSON using the ident-hash (uuid@version)."""
-    result = _get_content_json()
-
-    result = json.dumps(result)
-    resp = request.response
-    resp.status = "200 OK"
-    resp.content_type = 'application/json'
-    resp.body = result
-    return resp
-
-
-@view_config(route_name='content-html', request_method='GET')
-def get_content_html(request):
-    """Retrieve content as HTML using the ident-hash (uuid@version)."""
-    result = _get_content_json()
-
-    media_type = result['mediaType']
-    if media_type == COLLECTION_MIMETYPE:
-        content = tree_to_html(result['tree'])
-    else:
-        content = result['content']
-
-    resp = request.response
-    resp.status = "200 OK"
-    resp.content_type = 'application/xhtml+xml'
-    resp.body = content
-    return resp
-
-
-@view_config(route_name='content', request_method='GET')
-def get_content(request):
-    """Retrieve content using the ident-hash (uuid@version).
-
-    Depending on the HTTP_ACCEPT header return HTML or JSON.
-    """
-    if 'application/xhtml+xml' in request.headers.get('ACCEPT', ''):
-        return get_content_html(request)
-
-    else:
-        return get_content_json(request)
 
 
 @view_config(route_name='legacy-redirect', request_method='GET')
